requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pytest
httpx
google-generativeai
anyio
google-api-core